+++ packages/polythene-mithril-ripple/src/index.js
@@ -1,5 +1,5 @@
 import { _Ripple } from "polythene-core-ripple";
-import { cast, h, a, getRef, useState, useEffect } from "cyano-mithril";
+import { cast, h, a, getRef, useRef, useState, useEffect } from "cyano-mithril";
 
-export const Ripple = cast(_Ripple, { h, a, getRef, useState, useEffect });
+export const Ripple = cast(_Ripple, { h, a, getRef, useRef, useState, useEffect });
 Ripple["displayName"] = "Ripple";