+++ packages/polythene-react-ripple/src/index.js
@@ -1,4 +1,4 @@
 import { _Ripple } from "polythene-core-ripple";
-import { cast, h, a, getRef, useState, useEffect } from "cyano-react";
+import { cast, h, a, getRef, useRef, useState, useEffect } from "cyano-react";
 
-export const Ripple = cast(_Ripple, { h, a, getRef, useState, useEffect });
+export const Ripple = cast(_Ripple, { h, a, getRef, useRef, useState, useEffect });